Add GeofenceUtilities and move get\set functions to Controller
