+++ app/src/main/java/gr/teicm/informatics/selfdrivegps/Utilities/PermissionUtilities.java
@@ -97,8 +97,8 @@ public class PermissionUtilities extends Activity {
         locationRequest.setPriority(LocationRequest.PRIORITY_HIGH_ACCURACY);
         locationRequest.setInterval(30 * 1000);
         locationRequest.setFastestInterval(5 * 1000);
-        LocationSettingsRequest.Builder builder = new LocationSettingsRequest.Builder()
-                .addLocationRequest(locationRequest);
+
+        LocationSettingsRequest.Builder builder = new LocationSettingsRequest.Builder().addLocationRequest(locationRequest);
         builder.setAlwaysShow(true);
 
         PendingResult<LocationSettingsResult> result =
